Added CICD for Frontend and App Server
